requirements.txt
fastapi
uvicorn[standard]
algoliasearch
python-decouple
python-multipart
sentence-transformers